Implemented styling, added about page, more work with Spotify api.

+++ RipperProject/RipperApp/edit_song.py
@@ -1,5 +1,4 @@
 import os
-
 import eyed3
 
 
@@ -11,10 +10,8 @@ def editor(artist, title, album, genre, path_list, album_path):
     audio_file.tag.genre = genre
 
     # filename = os.path.basename(picture_path)
-
-    audio_file.tag.images.set(3, open(album_path, 'rb').read(), 'image/jpeg')
-    os.remove(album_path)
+    if album_path:
+        audio_file.tag.images.set(3, open(album_path, 'rb').read(), 'image/jpeg')
+        os.remove(album_path)
 
     audio_file.tag.save()
-
-    path_list = path_list[1:]
